requirements.txt
Flask==3.1.0
paho-mqtt==2.1.0 